forgot to save

+++ Obstacles/Guard.cs
@@ -1,5 +1,5 @@
 namespace Obstacles {
     class Guard : Obstacle{
-        public Guard(byte x, byte y) : base(x, y) {}
+        public Guard(int x, int y) : base(x, y) {}
     }
 }
